requirements.txt
absl-py
ipykernel
ipython
ipython-genutils
ipywidgets
jupyter-client
jupyter-core
jupyterlab
jupyterlab-server
Keras
kiwisolver
lazy-object-proxy
matplotlib
neptune-client
networkx
numpy
opencv-python
pandas
Pillow
PyWavelets
quantities
requests
scikit-learn
scipy
seaborn
sparse
sympy
termcolor
test-tube
torch
torchvision
tqdm
git+git://github.com/tihbe/python-ebdataset
fire==0.3.1
